Feature: self-update: rename module update -> self_update
