+++ src/self_updater/checker.rs
