+++ src/self_updater/downloader.rs
