+++ src/self_updater/environment.rs
